Fixed Live editor not closing Constellation if selected gameobject changed
